Add namespace metrics
